:truck: Rename: 꿀벌 이미지 이름 및 위치 변경
Rename: 꿀벌 이미지 이름 및 위치 변경
