+++ app/project/[projectId]/_components/Shield/WriteCommentShield.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Image from "next/image";
-import feedbeeIcon from "@/public/icons/feedbee2.svg";
+import feedbeeIcon from "@/public/beeIcons/yellowBee.svg";
 import fullStarIcon from "@/public/icons/fullStar.svg";
 import Button from "@/app/_components/Button/Button";
 import LoginModal from "@/app/_components/LoginModal/LoginModal";
