+++ app/project/[projectId]/comments/[ratingId]/not-found.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
-import angryFeedB from "@/public/icons/angryFeedB.png";
 import Button from "@/app/_components/Button/Button";
+import angryFeedB from "@/public/beeIcons/redBee.svg";
 
 function NotFound() {
   return (
